Move explorer files into a distinct component
