+++ src/api/implementation/cluster-explorer/versions.ts
@@ -1,7 +1,7 @@
 import * as v1 from "./v1";
 import { API } from "../../contract/api";
 import { versionUnknown, available } from "../apiutils";
-import { KubernetesExplorer } from "../../../explorer";
+import { KubernetesExplorer } from "../../../components/clusterexplorer/explorer";
 
 export function apiVersion(explorer: KubernetesExplorer, version: string): API<any> {
     switch (version) {
